pm-copilot/pyproject.toml
[project]
name = "pm-copilot"
version = "0.1.0"
description = "PM Research Copilot: from industry research to PRD, a full-stack product research workbench powered by multi-agent workflow"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "google-adk>=1.8.0",
    "litellm",
    "tavily-python",
    "chromadb",
    "langchain-text-splitters",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
    "jupyter>=1.1.1",
    "ruff>=0.4.6",
    "mypy>=1.19.0",
    "codespell>=2.4.1",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.hatch.build.targets.wheel]
packages = ["app", "frontend"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"